Add Authentication Authorization to login, getToken function module, Modify Mypage
