+++ app/src/pages/mypage/index.ts
@@ -1,6 +1,5 @@
 /** @format */
 
 import MyPage from './mypage/MyPage';
-import OtherPage from './other-page/OtherPage';
 
-export { MyPage, OtherPage };
+export { MyPage };
