+++ app/src/pages/mypage/mypage/MyPage.tsx
@@ -1,25 +1,9 @@
 /** @format */
 
-import { useState, useEffect } from 'react';
-import { css, cx } from '@emotion/css';
-import { keyframes } from '@emotion/react';
 import { useParams } from 'react-router-dom';
 import { get_user_info } from '../../../redux/mypage/reducer';
-import { color, radius, font, shadow } from '../../../styles';
-import Review from '../Review';
-import Slide from './Slide';
-import {
-  Img,
-  Poster,
-  NewPost,
-  Box,
-  Profile,
-  FocusBar,
-  Category,
-} from '../../../components';
 import Presenter from './Presenter';
-import axios from 'axios';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { RootState } from '../../../redux/root';
 import {
   get_user_specs,
@@ -27,6 +11,8 @@ import {
   get_user_helpme,
 } from '../../../redux/spec/reducer';
 import { useGetRequest } from '../../../redux/axios';
+import { decodeToken } from 'react-jwt';
+import getToken from '../../../utils/getToken';
 
 interface Props {
   [key: string]: any;
@@ -47,11 +33,17 @@ export default function MyPage({}: Props) {
       toHelp: get_user_tohelp,
       helpMe: get_user_helpme,
     },
-    userId = useParams().no,
+    TOKEN = getToken(),
+    tokenInfo = decodeToken<any>(TOKEN),
+    userNo = TOKEN !== '' && String(tokenInfo.userNo),
+    paramNo = useParams().no,
+    checkSelf = String(userNo === paramNo),
+    take: any = {
+      true: 3,
+      false: 4,
+    },
     userInfo = useSelector((state: RootState) => state.mypage.user.profile),
     posts = useSelector((state: RootState) => state.spec),
-    TOKEN =
-      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImVzdGFyZzFAaGFubWFpbC5uZXQiLCJ1c2VyTm8iOjUsImlzc3VlciI6Im1vZGVybi1hZ2lsZSIsImV4cGlyYXRpb24iOiIzNjAwMCIsImlhdCI6MTY1NjI0OTExNCwiZXhwIjoxNjU2Mjg1MTE0fQ.H8FvsMNRv40Z8sfiiTej2kV5i5H8Iyj0oU9DgmzgekE',
     text: { [key: string]: any } = {
       sir: '님',
       registerDate: '가입일 :',
@@ -67,26 +59,29 @@ export default function MyPage({}: Props) {
       rating: '총 평점',
     };
 
-  useGetRequest(`${SPEC}user=${userId}&take=6&page=1`, TOKEN, get_user_specs);
   useGetRequest(
-    `${BOARDS}user=${userId}&take=6&page=1${target[1]}`,
+    `${SPEC}user=${paramNo}&take=${take[checkSelf] * 2}&page=1`,
+    TOKEN,
+    get_user_specs
+  );
+  useGetRequest(
+    `${BOARDS}user=${paramNo}&take=${take[checkSelf] * 2}&page=1${target[1]}`,
     TOKEN,
     get_user_tohelp
   );
   useGetRequest(
-    `${BOARDS}user=${userId}&take=6&page=1${target[0]}`,
+    `${BOARDS}user=${paramNo}&take=${take[checkSelf] * 2}&page=1${target[0]}`,
     TOKEN,
     get_user_helpme
   );
-  useGetRequest(`profile/${userId}`, TOKEN, get_user_info);
-  console.log('해줄래요 게시글들 :>> ', posts.profileToHelp);
+  useGetRequest(`profile/${paramNo}`, TOKEN, get_user_info);
   return (
     <Presenter
       text={text}
       userInfo={userInfo && userInfo}
       posts={posts}
       actions={actions}
+      checkSelf={checkSelf}
     />
   );
 }
-// container
