+++ app/src/styles/layout/main/Main.tsx
@@ -14,6 +14,8 @@ export default function Main({ main, children }: Props): ReactElement {
     height: fit-content;
     overflow-x: visible;
   `;
+  const TOKEN = sessionStorage.getItem('userAccessToken');
+  // console.log('TOKEN :>> ', TOKEN);
 
   return <div className={cx(style)}>{children}</div>;
 }
