+++ app/src/utils/getToken.ts
@@ -1,0 +1,7 @@
+/** @format */
+
+const getToken = () => {
+  return sessionStorage.getItem('userAccessToken') || '';
+};
+
+export default getToken;
